+++ components/Link.js
@@ -4,8 +4,6 @@ export default (props) => {
   let href;
 
   if (props.as) {
-    console.log(props.as.indexOf('/blog/'));
-
     if (props.as.indexOf('/blog/') >= 0) {
       href = '/blog/[slug]';
     } else if (props.as.indexOf('/collection/item/' >= 0)) {