+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.1] - 2020-03-03
+
+- Add kids audio guide proxy
+
 ## [1.6.0] - 2019-11-05
 
-* Add this changelog (sorry, better late than never)
-* Add `off-the-shelf` proxy
+- Add this changelog (sorry, better late than never)
+- Add `off-the-shelf` proxy